fix: list hang, file bloat, compare OOM; add scheme descriptions
Major fixes:
1. list no longer hangs - lazy torch import
   - torch/engine/evaluate import inside functions only
   - list runs in ~60ms (was hanging on torch CUDA preload)
   - torch-heavy commands show loading hint first

2. FP8 file no longer bloats (5.5GB -> 3.1GB not 7.8GB)
   - RWKV .pth packs tensors into shared storage; quantizing a few
     weights kept the whole 5.7GB storage alive
   - compact_state() breaks shared storage before save
   - int8_symmetric emits real int8 dtype (was float32, 2x bloat)
   - int8_affine/int4 approximate weights restored to original dtype

3. compare no longer OOM-kills WSL on 2.9B model
   - load_state uses mmap lazy load
   - compare reloads per-scheme instead of cloning whole model

4. list shows per-scheme descriptions

+++ schemes.py
@@ -38,7 +38,7 @@ def _q_int8_sym(w: torch.Tensor, **kw):
     MAX = 127.0
     amax = w.abs().max().clamp(min=1e-12)
     scale = (amax / MAX).float()
-    w_q = (w.float() / scale).round().clamp(-128, 127)
+    w_q = (w.float() / scale).round().clamp(-128, 127).to(torch.int8)
     return w_q, scale
 
 def _q_int8_affine(w: torch.Tensor, **kw):
